% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(layers = 10)
}
\arguments{
\item{layers}{Desired number of output layers.}
}
\value{
A list of \code{sf} polygons.
}
\description{
This function can be used to generate randomised test data.
}
