% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_graph}
\alias{get_graph}
\title{Graphs}
\usage{
get_graph(ts_dat)
}
\arguments{
\item{ts_dat}{time series of \code{sf} objects.}
}
\value{
An \code{igraph} object representing all objects and their
connections.
}
\description{
Calculate the graphs connecting objects across time.
}
