% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-geometry-size-attributes.R
\name{getArea}
\alias{getArea}
\title{get area of polygon}
\usage{
getArea(polygon)
}
\arguments{
\item{polygon}{\code{sf} polygon}
}
\value{
area of the \code{sf} polygon in units of coordinate system
}
\description{
This function calculates the area of a \code{sf} polygon.
}
