% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-geometry-size-attributes.R
\name{getCompactness}
\alias{getCompactness}
\title{get compactness of polygon}
\usage{
getCompactness(polygon)
}
\arguments{
\item{polygon}{\code{sf} polygon}
}
\value{
compactness of the \code{sf} polygon in units of coordinate system
}
\description{
This function calculates the compactness of a \code{sf} polygon.
}
