% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-geometry-size-attributes.R
\name{getOrientation}
\alias{getOrientation}
\title{get orientation of polygon}
\usage{
getOrientation(polygon)
}
\arguments{
\item{polygon}{\code{sf} polygon}
}
\value{
orientation of the \code{sf} polygon in degree
}
\description{
This function calculates the orientation of a \code{sf} polygon.
}
