% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-geometry-size-attributes.R
\name{getPerimeter}
\alias{getPerimeter}
\title{get perimeter of polygon}
\usage{
getPerimeter(polygon)
}
\arguments{
\item{polygon}{\code{sf} polygon}
}
\value{
perimeter of the \code{sf} polygon in units of coordinate system
}
\description{
This function calculates the perimeter of a \code{sf} polygon.
}
