% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersections}
\alias{intersections}
\title{Intersections}
\usage{
intersections(l1, l2)
}
\arguments{
\item{l1}{\code{sfc} polygon layer at time 1.}

\item{l2}{\code{sfc} polygon layer at time 2.}
}
\value{
A \code{sfc} polygon layer.
}
\description{
Determine which objects in one layer intersect objects of another layer.
}
