% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{qualitative_cos}
\alias{qualitative_cos}
\title{Qualitative change of support}
\usage{
qualitative_cos(dat)
}
\arguments{
\item{dat}{\code{igraph} object representing all objects and layers.}
}
\value{
Map of qualitative COS
}
\description{
Calculate the qualitative change of suppot that connected nodes in a graph
undergo. The support can be one of the following: stable, split, merge,
anihilation, creation.
}
